requirements.txt
openai>=1.0.0
sounddevice>=0.4.6
scipy>=1.10.0
pynput>=1.7.6
python-dotenv>=1.0.0
numpy>=1.24.0
